moved protocol functions to new source file
