+++ gapii/client/BUILD.bazel
@@ -22,6 +22,7 @@ go_library(
         "doc.go",
         "header.go",
         "jdwp_loader.go",
+        "protocol.go",
     ],
     importpath = "github.com/google/gapid/gapii/client",
     visibility = ["//visibility:public"],
